requirements.txt
Fabric==1.4.3
Flask==0.9
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.8.3
bcdoc==0.12.0
boto==2.34.0
botocore==0.31.0
clan==0.1.3
colorama==0.2.5
copytext==0.1.8
cssmin==0.2.0
docutils==0.11
facebook-sdk==0.4.0
gunicorn==19.1.1
httplib2==0.9
jmespath==0.2.1
nose==1.2.1
odict==1.5.1
openpyxl==2.1.4
ply==3.4
pyasn1==0.1.7
pycrypto==2.6
python-dateutil==2.2
requests==2.5.0
six==1.5.2
slimit==0.7.4
smartypants==1.8.6
ssh==1.7.14
termcolor==1.1.0
twitter==1.14.3
wheel==0.23.0
wsgiref==0.1.2
